--- packages/roosterjs-content-model-plugins/lib/autoFormat/horizontalLine/HorizontalLineTriggerCharacter.ts
@@ -1,4 +1,0 @@
-/**
- * @internal
- */
-export type HorizontalLineTriggerCharacter = '-' | '=' | '_' | '*' | '~' | '#';
